third_party/blink/web_tests/external/wpt/html/semantics/forms/the-input-element/input-type-change-value-expected.txt

This is a testharness.js-based test.
[FAIL] Without focus
  assert_equals: Value dirty flag should remain false expected "" but got "#000000"
[FAIL] With focus
  assert_equals: Value dirty flag should remain false expected "" but got "#000000"
Harness: the test ran to completion.
